Support change events
Can now respond to change events like one can for add and remove. The
magic prefix is "#" and such commands will only be run with this explicit
prefix (ie not with "*")

To be compatible with mdev/smdev, the device addition/removal code isn't
executed for change events. Only external commands can optionally be
run.